+++ src/app/cart/page.tsx
@@ -1,13 +1,5 @@
-import BreadCrumb from "@/components/bread-crumb";
-import CartContent from "./_components/cart-content";
-import RecentlyViewed from "./_components/recently-viewed";
+import PageContent from "./_components/page-content";
 
 export default function Cart() {
-  return (
-    <>
-      <BreadCrumb page="Cart" key="cart" />
-      <CartContent />
-      <RecentlyViewed />
-    </>
-  );
+  return <PageContent />;
 }
